Update to 2.2.15
This is mainly a hotfix to address all of the issues that came with 2.2.14, so there’s not much to write home about, except that now you can play the game as intended again.

Additionally, we have chosen to revert Pipe Towers Zone to its previous version in order to resolve some asset crediting issues that sprang up after launch – for this and other reasons, 2.2.14 will no longer be available for download.